Specification/resource code unit refactoring.
